services/producer/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0

